requirements.txt
streamlit
st_pages
st-gsheets-connection
pandas
pillow
requests
plotly
